+++ packages/components/organisms/f-header/src/components/_tests/Header.test.js
@@ -5,7 +5,8 @@ const $style = {
     'c-header': 'c-header',
     'c-header-container': 'c-header-container',
     'c-header--transparent': 'c-header--transparent',
-    'c-header--gradient': 'c-header--gradient'
+    'c-header--gradient': 'c-header--gradient',
+    'c-header--disappearingWhite': 'c-header--disappearingWhite'
 };
 
 describe('Header', () => {