+++ packages/components/pages/f-account-info/test-utils/component-objects/f-account-info-error.component.js
@@ -1,0 +1,12 @@
+import Page from '@justeat/f-wdio-utils';
+
+class AccountInfoErrorCard extends Page {
+    constructor () {
+        super('page', 'account-info-component');
+    }
+
+    // Overriding base implementation
+    get component () { return $('[data-test-id="account-info-error-card"]'); }
+}
+
+export default new AccountInfoErrorCard();