+++ packages/components/pages/f-checkout/test/visual/f-checkout-guest.visual.desktop.spec.js
@@ -1,6 +1,4 @@
-const Checkout = require('../../test-utils/component-objects/f-checkout.component');
-
-let checkout = new Checkout();
+import Checkout from '../../test-utils/component-objects/f-checkout.component';
 
 const illegalMobileNumber = '123';
 
@@ -13,11 +11,8 @@ describe('f-checkout - Collection - Guest - Desktop Visual Tests', () => {
     };
 
     beforeEach(async () => {
-        // Arrange
-        checkout = new Checkout();
-
         // Act
-        await checkout.load({
+        await Checkout.load({
             ...checkoutInfo
         });
     });
@@ -29,7 +24,7 @@ describe('f-checkout - Collection - Guest - Desktop Visual Tests', () => {
 
     it('should display the mandatory error messages', async () => {
         // Act
-        await checkout.goToPayment();
+        await Checkout.goToPayment();
 
         // Assert
         await browser.percyScreenshot('f-checkout - Collection - Guest - Mandatory Errors State', 'desktop');
@@ -37,8 +32,8 @@ describe('f-checkout - Collection - Guest - Desktop Visual Tests', () => {
 
     it('should display the illegal mobile number error message', async () => {
         // Act
-        await checkout.setFieldValue('mobileNumber', illegalMobileNumber);
-        await checkout.goToPayment();
+        await Checkout.setFieldValue('mobileNumber', illegalMobileNumber);
+        await Checkout.goToPayment();
 
         // Assert
         await browser.percyScreenshot('f-checkout - Collection - Guest - Illegal Mobile Number Error State', 'desktop');
@@ -53,18 +48,18 @@ describe('f-checkout - Collection - Guest - Desktop Visual Tests', () => {
             mobileNumber: '07234567890'
         };
 
-        checkout.inputFieldValues = customerInfo;
+        Checkout.inputFieldValues = customerInfo;
 
         // Act
-        await checkout.load({
+        await Checkout.load({
             ...checkoutInfo,
             isAsapAvailable: true,
             placeOrderError: 'duplicate'
         });
 
-        await checkout.setFieldValues();
+        await Checkout.setFieldValues();
         await browser.pause(200);
-        await checkout.goToPayment();
+        await Checkout.goToPayment();
 
         // Assert
         await browser.percyScreenshot('f-checkout - Collection - Guest - "Duplicate Order Warning" Modal', 'desktop');
@@ -75,7 +70,7 @@ describe('f-checkout - Collection - Guest - Desktop Visual Tests', () => {
         const emailAddress = '@jazz.man@tunetown.com';
 
         // Act
-        await checkout.setFieldValue('emailAddress', emailAddress);
+        await Checkout.setFieldValue('emailAddress', emailAddress);
         await browser.pause(200);
         await browser.keys('Tab');
 
@@ -86,11 +81,8 @@ describe('f-checkout - Collection - Guest - Desktop Visual Tests', () => {
 
 describe('f-checkout - Collection - Guest - isAsapAvailable: false Desktop Visual Tests', () => {
     beforeEach(async () => {
-        // Arrange
-        checkout = new Checkout();
-
         // Act
-        await checkout.load({
+        await Checkout.load({
             serviceType: 'collection',
             isLoggedIn: false,
             isAsapAvailable: false,
@@ -113,11 +105,8 @@ describe('f-checkout - Delivery - Guest - Desktop Visual Tests', () => {
     };
 
     beforeEach(async () => {
-        // Arrange
-        checkout = new Checkout();
-
         // Act
-        await checkout.load({
+        await Checkout.load({
             ...checkoutInfo
         });
     });
@@ -129,7 +118,7 @@ describe('f-checkout - Delivery - Guest - Desktop Visual Tests', () => {
 
     it('should display the delivery f-checkout guest mandatory error messages', async () => {
         // Act
-        await checkout.goToPayment();
+        await Checkout.goToPayment();
 
         // Assert
         await browser.percyScreenshot('f-checkout - Delivery - Guest - Mandatory Errors State', 'desktop');
@@ -137,8 +126,8 @@ describe('f-checkout - Delivery - Guest - Desktop Visual Tests', () => {
 
     it('should display the illegal mobile number error message', async () => {
         // Act
-        await checkout.setFieldValue('mobileNumber', illegalMobileNumber);
-        await checkout.goToPayment();
+        await Checkout.setFieldValue('mobileNumber', illegalMobileNumber);
+        await Checkout.goToPayment();
 
         // Assert
         await browser.percyScreenshot('f-checkout - Delivery - Guest - Illegal Mobile Number Error State', 'desktop');
@@ -146,8 +135,6 @@ describe('f-checkout - Delivery - Guest - Desktop Visual Tests', () => {
 
     it('should display the "Duplicate Order Warning" modal', async () => {
         // Arrange
-        checkout = new Checkout();
-
         const customerInfo = {
             firstName: 'Jerry',
             lastName: 'Jazzman',
@@ -158,16 +145,16 @@ describe('f-checkout - Delivery - Guest - Desktop Visual Tests', () => {
             addressPostcode: 'W1D 4HT'
         };
 
-        checkout.inputFieldValues = customerInfo;
+        Checkout.inputFieldValues = customerInfo;
 
         // Act
-        await checkout.load({
+        await Checkout.load({
             ...checkoutInfo,
             placeOrderError: 'duplicate'
         });
-        await checkout.setFieldValues();
+        await Checkout.setFieldValues();
         await browser.pause(200);
-        await checkout.goToPayment();
+        await Checkout.goToPayment();
 
         // Assert
         await browser.percyScreenshot('f-checkout - Delivery - Guest - "Duplicate Order Warning" Modal', 'desktop');
@@ -178,7 +165,7 @@ describe('f-checkout - Delivery - Guest - Desktop Visual Tests', () => {
         const emailAddress = '@jazz.man@tunetown.com';
 
         // Act
-        await checkout.setFieldValue('emailAddress', emailAddress);
+        await Checkout.setFieldValue('emailAddress', emailAddress);
         await browser.pause(200);
         await browser.keys('Tab');
 
@@ -189,11 +176,8 @@ describe('f-checkout - Delivery - Guest - Desktop Visual Tests', () => {
 
 describe('f-checkout - Delivery - Guest - isAsapAvailable: false Desktop Visual Tests', () => {
     beforeEach(async () => {
-        // Arrange
-        checkout = new Checkout();
-
         // Act
-        await checkout.load({
+        await Checkout.load({
             serviceType: 'delivery',
             isLoggedIn: false,
             isAsapAvailable: false,
@@ -216,11 +200,8 @@ describe('f-checkout - Dine In - Guest - Desktop Visual Tests', () => {
     };
 
     beforeEach(async () => {
-        // Arrange
-        checkout = new Checkout();
-
         // Act
-        await checkout.load({
+        await Checkout.load({
             ...checkoutInfo
         });
     });
@@ -232,8 +213,8 @@ describe('f-checkout - Dine In - Guest - Desktop Visual Tests', () => {
 
     it('should display the mandatory error messages', async () => {
         // Act
-        await checkout.clearBlurField('tableIdentifier');
-        await checkout.goToPayment();
+        await Checkout.clearBlurField('tableIdentifier');
+        await Checkout.goToPayment();
 
         // Assert
         await browser.percyScreenshot('f-checkout - Dine In - Guest - Manadatory Errors', 'desktop');
@@ -241,8 +222,8 @@ describe('f-checkout - Dine In - Guest - Desktop Visual Tests', () => {
 
     it('should display the illegal mobile number error message', async () => {
         // Act
-        await checkout.setFieldValue('mobileNumber', illegalMobileNumber);
-        await checkout.goToPayment();
+        await Checkout.setFieldValue('mobileNumber', illegalMobileNumber);
+        await Checkout.goToPayment();
 
         // Assert
         await browser.percyScreenshot('f-checkout - Dine In - Guest - Illegal Mobile Number Error State', 'desktop');
@@ -250,8 +231,6 @@ describe('f-checkout - Dine In - Guest - Desktop Visual Tests', () => {
 
     it('should display the "Duplicate Order Warning" modal', async () => {
         // Arrange
-        checkout = new Checkout();
-
         const customerInfo = {
             firstName: 'Jerry',
             lastName: 'Jazzman',
@@ -260,16 +239,16 @@ describe('f-checkout - Dine In - Guest - Desktop Visual Tests', () => {
             tableIdentifier: '10'
         };
 
-        checkout.inputFieldValues = customerInfo;
+        Checkout.inputFieldValues = customerInfo;
 
         // Act
-        await checkout.load({
+        await Checkout.load({
             ...checkoutInfo,
             placeOrderError: 'duplicate'
         });
-        await checkout.setFieldValues();
+        await Checkout.setFieldValues();
         await browser.pause(200);
-        await checkout.goToPayment();
+        await Checkout.goToPayment();
 
         // Assert
         await browser.percyScreenshot('f-checkout - Dine in - Guest - "Duplicate Order Warning" Modal', 'desktop');