+++ packages/components/pages/f-takeawaypay-activation/test-utils/component-objects/f-takeawaypayActivation-selectors.js
@@ -1,10 +1,23 @@
-exports.TAKEAWAYPAYACTIVATION_COMPONENT = '[data-test-id="takeawaypay-activation-component"]';
-exports.LOGGEDIN_COMPONENT = '[data-test-id="activation-logged-in-component"]';
-exports.ERROR_COMPONENT = '[data-test-id="activation-failed-component"]';
-exports.ACTIVATION_SUCCESS_COMPONENT = '[data-test-id="activation-successful-component"]';
-exports.ACTIVATE_TAKEAWAYPAY_BUTTON = '[data-test-id="takeawaypay-activation-activate-button"]';
-exports.UNAUTHENTICATED_LOGIN_BUTTON = '[data-test-id="takeawaypay-activation-not-loggedin-login-link"]';
-exports.UNAUTHENTICATED_REGISTER_BUTTON = '[data-test-id="takeawaypay-activation-not-loggedin-registration-link"]';
-exports.AUTHENTICATED_LOGIN_BUTTON = '[data-test-id="takeawaypay-activation-loggedin-login-link"]';
-exports.AUTHENTICATED_REGISTER_BUTTON = '[data-test-id="takeawaypay-activation-loggedin-registration-link"]';
-exports.START_ORDERING_BUTTON = '[data-test-id="takeawaypay-activation-home-link"]';
+const TAKEAWAYPAYACTIVATION_COMPONENT = '[data-test-id="takeawaypay-activation-component"]';
+const LOGGEDIN_COMPONENT = '[data-test-id="activation-logged-in-component"]';
+const ERROR_COMPONENT = '[data-test-id="activation-failed-component"]';
+const ACTIVATION_SUCCESS_COMPONENT = '[data-test-id="activation-successful-component"]';
+const ACTIVATE_TAKEAWAYPAY_BUTTON = '[data-test-id="takeawaypay-activation-activate-button"]';
+const UNAUTHENTICATED_LOGIN_BUTTON = '[data-test-id="takeawaypay-activation-not-loggedin-login-link"]';
+const UNAUTHENTICATED_REGISTER_BUTTON = '[data-test-id="takeawaypay-activation-not-loggedin-registration-link"]';
+const AUTHENTICATED_LOGIN_BUTTON = '[data-test-id="takeawaypay-activation-loggedin-login-link"]';
+const AUTHENTICATED_REGISTER_BUTTON = '[data-test-id="takeawaypay-activation-loggedin-registration-link"]';
+const START_ORDERING_BUTTON = '[data-test-id="takeawaypay-activation-home-link"]';
+
+export {
+    TAKEAWAYPAYACTIVATION_COMPONENT,
+    LOGGEDIN_COMPONENT,
+    ERROR_COMPONENT,
+    ACTIVATION_SUCCESS_COMPONENT,
+    ACTIVATE_TAKEAWAYPAY_BUTTON,
+    UNAUTHENTICATED_LOGIN_BUTTON,
+    UNAUTHENTICATED_REGISTER_BUTTON,
+    AUTHENTICATED_LOGIN_BUTTON,
+    AUTHENTICATED_REGISTER_BUTTON,
+    START_ORDERING_BUTTON
+};