+++ packages/components/organisms/f-header/CHANGELOG.md
@@ -2,6 +2,14 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v10.2.0
+-----------------------------
+*July 1, 2022*
+
+### Added
+- Theme "disappearingWhite" to support header on Search
+
+
 v10.1.0
 ------------------------------
 *July 7, 2022*