+++ packages/components/pages/f-contact-preferences/CHANGELOG.md
@@ -4,12 +4,21 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
-v1.1.0
+v1.2.0
 -----------------------------
-*July 14, 2022*
+*July 15, 2022*
 
 ### Added
 - Peer dependencies for all fozzie components.
+
+
+v1.1.0
+------------------------------
+*July 13, 2022*
+
+### Changed
+- Update f-wdio-utils to v1.1.0 to use new Axe Core implementation.
+- Accessibility tests to be async.
 
 
 v1.0.0