+++ packages/components/pages/f-registration/CHANGELOG.md
@@ -4,12 +4,32 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
-v2.5.0
+v3.2.0
 ------------------------------
-*July 8, 2021*
+*July 15, 2021*
+
+### Added
+- Peer dependencies for fozzie components.
+
+
+v3.1.0
+------------------------------
+*July 13, 2022*
+
+### Changed
+- Update f-wdio-utils to v1.1.0 to use new Axe Core implementation.
+- Accessibility tests to be async.
 
 ### Changed
 - Refactor WebDriverIO tests to use async in order to support Node 16 using `codemod` utility.
+
+
+v3.0.0
+------------------------------
+*July 13, 2022*
+
+### Changed
+- **breaking changes** Update to `@use` and `@forward` SASS syntax.
 
 
 v2.4.2