+++ packages/components/pages/f-takeawaypay-activation/CHANGELOG.md
@@ -4,12 +4,21 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
-v2.14.0
+v2.15.0
 ------------------------------
-*July 8, 2021*
+*July 15, 2021*
 
 ### Added
-- Peer dependencies for all fozzie components.
+- Peer dependencies for fozzie components.
+
+
+v2.14.0
+------------------------------
+*July 13, 2022*
+
+### Changed
+- Update f-wdio-utils to v1.1.0 to use new Axe Core implementation.
+- Accessibility tests to be async.
 
 ### Changed
 - Refactor WebDriverIO tests to use async in order to support Node 16 using `codemod` utility.